requirements.txt
pyowm~=3.2.0
schedule~=1.1.0
python-dotenv~=0.19.2
pyTelegramBotAPI~=4.3.1
pip~=21.3.1
geojson~=2.5.0
requests~=2.27.1
certifi~=2021.10.8
urllib3~=1.26.8
idna~=3.3
setuptools~=58.1.0